+++ go/models/player.go
@@ -8,8 +8,8 @@ type Player struct {
 	Attack    int
 	Defense   int
 	Speed     float64
-	Crit      int
-	Dodge     int
-	Block     int
+	Crit      float64
+	Dodge     float64
+	Block     float64
 	Ore       int
 }
